Base the syntax table on c-mode-syntax-table.
This is much closer to Faust's lexical syntax, so that we only have to fix up
a few characters which have special roles in C but are just ordinary
punctuation in Faust. This also fixes some issues with symbol constituents in
the standard syntax table which get in the way of auto-completion.